QA-7819 removed JIRA logic from HQ conftest and added to common fixture
