+++ HQSmokeTests/testCases/conftest.py
@@ -12,8 +12,6 @@ from datetime import datetime
 
 global driver
 
-failed_items = []
-
 @pytest.fixture(scope="session")
 def environment_settings_hq():
     """Load settings from os.environ
@@ -97,42 +95,3 @@ def pytest_terminal_summary(terminalreporter, exitstatus, config):
         f.write(f'ERROR={len(error)}\n')
         f.write(f'SKIPPED={len(skipped)}\n')
         f.write(f'XFAIL={len(xfail)}\n')
-
-def pytest_configure(config):
-    global pytest_html
-    pytest_html = config.pluginmanager.getplugin('html')
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_runtest_makereport(item, call):
-    outcome = yield
-    report = outcome.get_result()
-
-    # Call existing screenshot hook from fixtures if it exists
-    try:
-        from common_utilities.fixtures import pytest_runtest_makereport as screenshot_hook
-        # Simulate calling the other makereport logic manually
-        screenshot_hook(item)
-    except (ImportError, AttributeError):
-        pass
-
-    if report.when == "call" and report.failed:
-        failed_items.append(item)
-
-
-def pytest_sessionfinish(session, exitstatus):
-    if not failed_items:
-        return
-
-    lines = []
-    for item in failed_items:
-        try:
-            doc = item.function.__doc__ or "No reproduction steps provided."
-        except AttributeError:
-            doc = "No docstring available (non-function test case)"
-        lines.append(f"Test: {item.nodeid}\nRepro Steps:\n{doc.strip()}\n\n---")
-
-    with open("jira_ticket_body.txt", "w", encoding="utf-8") as f:
-        f.write(f"🔥 Automated Failure Report - {datetime.now().strftime('%Y-%m-%d %H:%M')}\n\n")
-        f.write("\n".join(lines) if lines else "✅ All tests passed.")
-
-
