+++ common_utilities/fixtures.py
@@ -4,12 +4,14 @@ from py.xml import html
 from selenium import webdriver
 from common_utilities.path_settings import PathSettings
 from common_utilities.hq_login.login_page import LoginPage
+import base64
+from datetime import datetime
 
 
 """"This file provides fixture functions for driver initialization"""
 
 global driver
-
+failed_items = []
 
 @pytest.fixture(scope="module", autouse=True)
 def driver(settings, browser):
@@ -111,24 +113,50 @@ def pytest_html_results_table_row(report, cells):
     cells.pop()
 
 
+def _capture_screenshot(driver):
+    return base64.b64encode(driver.get_screenshot_as_png()).decode('utf-8')
+
 @pytest.hookimpl(hookwrapper=True)
 def pytest_runtest_makereport(item):
-    pytest_html = item.config.pluginmanager.getplugin("html")
     outcome = yield
     report = outcome.get_result()
+    pytest_html = item.config.pluginmanager.getplugin("html")
+
     tags = ", ".join([m.name for m in item.iter_markers() if m.name != 'run'])
     extra = getattr(report, 'extra', [])
-    if report.when == "call" or report.when == "teardown":
+
+    if report.when in ("call", "teardown"):
         xfail = hasattr(report, 'wasxfail')
         if (report.skipped and xfail) or (report.failed and not xfail):
-            print("reports skipped or failed")
-            file_name = report.nodeid.replace("::", "_") + ".png"
-            screen_img = _capture_screenshot(item.funcargs["driver"])
-            if file_name:
-                html = '<div><img src="data:image/png;base64,%s" alt="screenshot" style="width:600px;height:300px;" ' \
-                       'onclick="window.open(this.src)" align="right"/></div>' % screen_img
+            driver = item.funcargs.get("driver", None)
+            if driver:
+                screen_img = _capture_screenshot(driver)
+                html = (
+                    '<div><img src="data:image/png;base64,%s" alt="screenshot" '
+                    'style="width:600px;height:300px;" onclick="window.open(this.src)" '
+                    'align="right"/></div>' % screen_img
+                )
                 extra.append(pytest_html.extras.html(html))
-        report.extra = extra
-        report.tags = tags
-def _capture_screenshot(web_driver):
-    return web_driver.get_screenshot_as_base64()
+
+        if report.failed:
+            failed_items.append(item)
+
+    report.extra = extra
+    report.tags = tags
+
+
+def pytest_sessionfinish(session, exitstatus):
+    if not failed_items:
+        return
+
+    lines = []
+    for item in failed_items:
+        try:
+            doc = item.function.__doc__ or "No reproduction steps provided."
+        except AttributeError:
+            doc = "No docstring available (non-function test case)"
+        lines.append(f"Test: {item.nodeid}\nRepro Steps:\n{doc.strip()}\n\n---")
+
+    with open("jira_ticket_body.txt", "w", encoding="utf-8") as f:
+        f.write(f"🔥 Automated Failure Report - {datetime.now().strftime('%Y-%m-%d %H:%M')}\n\n")
+        f.write("\n".join(lines) if lines else "✅ All tests passed.")
